restructuring a bit with less regex
